backend/pyproject.toml
[project]
name = "redux-unifi-portal-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic-settings>=2.4",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.2",
  "httpx>=0.27",
  "redis>=5.0",
  "celery>=5.4",
  "authlib>=1.3",
  "python-multipart>=0.0.9",
  "structlog>=24.2",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt<4.0.0",
  "itsdangerous>=2.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
  "mypy>=1.10",
  "types-redis>=4.6.0.20241004",
]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
